backend/requirements.txt
fastapi[standard]
sqlmodel
asyncpg
pydantic-settings
alembic
python-dotenv
passlib[argon2]
python-jose[cryptography]
fastapi-mail
requests
Authlib
msgpack
fernet
google-auth
cloudinary
confluent-kafka
tenacity